+++ tests/test_graph.py
@@ -70,7 +70,7 @@ def test_complete_graph():
     expected_adjacency_matrix = sp.sparse.csr_matrix([[0, 1, 1, 1], [1, 0, 1, 1], [1, 1, 0, 1], [1, 1, 1, 0]])
 
     assert graph.number_of_vertices() == 4
-    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff = (graph.adjacency_matrix - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0
 
@@ -92,7 +92,7 @@ def test_cycle_graph():
                                                       [1, 0, 0, 1, 0]])
 
     assert graph.number_of_vertices() == 5
-    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff = (graph.adjacency_matrix - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0
 
@@ -114,7 +114,7 @@ def test_star_graph():
                                                       [1, 0, 0, 0, 0]])
 
     assert graph.number_of_vertices() == 5
-    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff = (graph.adjacency_matrix - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0
 
@@ -136,9 +136,11 @@ def test_path_graph():
                                                       [0, 0, 0, 1, 0]])
 
     assert graph.number_of_vertices() == 5
-    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff = (graph.adjacency_matrix - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0
+
+    assert graph.total_volume() == 8
 
     # Make sure we can't do something stupid
     with pytest.raises(ValueError):
@@ -182,7 +184,7 @@ def test_symmetry():
     big_graph = sgtl.random.ssbm(1000, 5, 0.8, 0.2)
 
     # Check that all of the graph matrices are truly symmetric
-    assert np.allclose(big_graph.adj_mat.toarray(), big_graph.adj_mat.toarray().T)
+    assert np.allclose(big_graph.adjacency_matrix.toarray(), big_graph.adjacency_matrix.toarray().T)
 
     lap_mat = big_graph.normalised_laplacian_matrix()
     lap_mat_dense = lap_mat.toarray()
@@ -231,7 +233,7 @@ def test_num_edges():
     # Check the number of edges in the graph
     assert graph.number_of_vertices() == 10
     assert graph.number_of_edges() == 21
-    assert graph.total_volume() == 21
+    assert graph.total_volume() == 42
 
     # Now create a weighted graph and check the number of edges method.
     adjacency_matrix = scipy.sparse.csr_matrix([[0, 2, 0, 1],
@@ -241,7 +243,7 @@ def test_num_edges():
     graph = sgtl.Graph(adjacency_matrix)
     assert graph.number_of_vertices() == 4
     assert graph.number_of_edges() == 4
-    assert graph.total_volume() == 7
+    assert graph.total_volume() == 14
 
     # Test the number of edges for a graph with self-loops
     adjacency_matrix = scipy.sparse.csr_matrix([[0, 2, 0, 1],
@@ -251,7 +253,7 @@ def test_num_edges():
     graph = sgtl.Graph(adjacency_matrix)
     assert graph.number_of_vertices() == 4
     assert graph.number_of_edges() == 5
-    assert graph.total_volume() == 9
+    assert graph.total_volume() == 16
 
     # Add more self-loops to a graph
     adjacency_matrix = scipy.sparse.csr_matrix([[0, 1, 1, 1, 1, 0, 0, 0, 0, 0],
@@ -268,7 +270,7 @@ def test_num_edges():
     graph = sgtl.Graph(adjacency_matrix)
     assert graph.number_of_vertices() == 10
     assert graph.number_of_edges() == 24
-    assert graph.total_volume() == 26.5
+    assert graph.total_volume() == 49
 
 
 def test_float_weights():
@@ -280,7 +282,7 @@ def test_float_weights():
     graph = sgtl.Graph(adjacency_matrix)
     assert graph.number_of_vertices() == 4
     assert graph.number_of_edges() == 5
-    assert graph.total_volume() == pytest.approx(9.99)
+    assert graph.total_volume() == pytest.approx(17.38)
 
     # Check the weight between vertex sets
     assert graph.weight([0], [1]) == pytest.approx(2.2)
@@ -306,7 +308,7 @@ def test_networkx():
                                                       [0, 0, 0, 0, 1, 1, 0, 1, 0],
                                                       [0, 0, 0, 0, 1, 1, 1, 0, 0],
                                                       [0, 0, 0, 1, 1, 0, 0, 0, 0]])
-    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff = (graph.adjacency_matrix - expected_adjacency_matrix)
     adj_mat_diff.eliminate_zeros()
     assert adj_mat_diff.nnz == 0